Fixed the calculation of hour, minutes, seconds and milliseconds when shift is a float number
Fixed the calculation of hour, minutes, seconds and milliseconds when shift is a float number

This fixed the error when using float number in line: 
out += '%s:%s:%s,%s --> %s:%s:%s,%s\n'

Also added a override option. If the third input parameter is "override" then save the result in the original file.